25.04.09.(수) 11:50

+++ src/main/java/com/solomarket/dto/FileDto.java
@@ -15,8 +15,8 @@ import lombok.Data;
  */
 @Data
 public class FileDto {
-    private Long fileNo;
-    private Long productNo;
+    private int fileNo;
+    private int productNo;
     private String fileName;
     private String fileUrl;
     private String fileType;
